DEMOS-512 remove duplicated code
