+++ client/src/components/table/tables/AmendmentTable.tsx
@@ -1,18 +1,5 @@
 import * as React from "react";
-
-import {
-  ChevronRightIcon,
-  SuccessIcon,
-} from "components/icons";
-import { ReviewIcon } from "components/icons/Action/ReviewIcon";
-
-import {
-  ExpandedState,
-  getCoreRowModel,
-  getExpandedRowModel,
-  useReactTable,
-} from "@tanstack/react-table";
-
+import { ExpandableTable } from "components/table/ExpandableTable";
 import {
   AmendmentColumns,
   RawAmendment,
@@ -25,101 +12,11 @@ interface AmendmentTableProps {
 }
 
 export function AmendmentTable({ data, initiallyExpandedId }: AmendmentTableProps) {
-  const [expanded, setExpanded] = React.useState<ExpandedState>(() => {
-    return initiallyExpandedId ? { [initiallyExpandedId]: true } : {};
-  });
-
-  const handleExpandedChange = React.useCallback(
-    (updater: ExpandedState | ((prev: ExpandedState) => ExpandedState)) => {
-      setExpanded((prev) => {
-        return typeof updater === "function" ? updater(prev) : updater;
-      });
-    },
-    []
-  );
-
-  const table = useReactTable({
-    data,
-    columns: AmendmentColumns,
-    state: { expanded },
-    onExpandedChange: handleExpandedChange,
-    getCoreRowModel: getCoreRowModel(),
-    getExpandedRowModel: getExpandedRowModel(),
-    getRowCanExpand: () => true,
-    getRowId: (row) => row.id,
-  });
-
   return (
-    <div className="w-full">
-      <div className="flex flex-col gap-2">
-        {table.getRowModel().rows.map((row) => {
-          const { title, effectiveDate, status } = row.original;
-          const isExpanded = row.getIsExpanded();
-
-          return (
-            <div key={row.id} className="border rounded px-4 py-2 bg-white">
-              <div
-                onClick={() => row.toggleExpanded()}
-                className="grid grid-cols-[auto_1fr_auto_auto] items-center gap-4 cursor-pointer"
-              >
-                <div className="text-sm font-bold text-blue-900">{title}</div>
-
-                <div className="h-1" />
-
-                <div>{renderStatus(status)}</div>
-
-                <div className="flex items-center gap-3">
-                  <span className="text-sm text-gray-800">{formatDate(effectiveDate)}</span>
-                  <ChevronRightIcon
-                    className={`w-[1.4rem] h-[1.4rem] text-[var(--color-action)] transform transition-transform duration-200 ${isExpanded ? "rotate-90" : "rotate-0"}`}
-                  />
-                </div>
-              </div>
-              {isExpanded && (
-                <div className="mt-2 px-2 py-2 bg-gray-100 text-sm italic text-gray-600 rounded-sm">
-                  Expanded details coming soon.
-                </div>
-              )}
-            </div>
-          );
-        })}
-      </div>
-    </div>
+    <ExpandableTable
+      data={data}
+      columns={AmendmentColumns}
+      initiallyExpandedId={initiallyExpandedId}
+    />
   );
 }
-
-const formatDate = (iso: string) => {
-  const [yyyy, mm, dd] = iso.split("-");
-  return `${mm}/${dd}/${yyyy}`;
-};
-
-const renderStatus = (status: string) => {
-  const baseStyle = "flex items-center gap-1 text-sm";
-  switch (status) {
-    case "Under Review":
-      return (
-        <div className={`${baseStyle} text-left`}>
-          <ReviewIcon className="w-[1.4rem] h-[1.4rem] text-yellow-500" />
-          {status}
-        </div>
-      );
-    case "Approved":
-      return (
-        <div className={`${baseStyle} text-left`}>
-          <SuccessIcon className="w-[1.4rem] h-[1.4rem]" />
-          {status}
-        </div>
-      );
-    case "Draft":
-      return (
-        <div className={`${baseStyle} text-left`}>
-          <SuccessIcon className="w-[1.4rem] h-[1.4rem]" />
-          {status}
-        </div>
-      );
-    default:
-      return (
-        <span className="text-sm text-gray-700">{status}</span>
-      );
-  }
-};
